changed integers to be unsigned

+++ community/c++-algo/include/GameLib/src/gameState.cpp
@@ -48,7 +48,7 @@ namespace terminal {
     /// Fills in the appropriate information for a player by reference.
     /// @param player The player to parse the stats for (by reference).
     /// @param stats A Json array containing the stats for the player.
-    void GameState::parsePlayerStats(Player& player, int id, Json::array stats) {
+    void GameState::parsePlayerStats(Player& player, unsigned int id, Json::array stats) {
         player.health = stats.at(0).int_value();
         player.cores = stats.at(1).number_value();
         player.bits = stats.at(2).number_value();
@@ -68,8 +68,8 @@ namespace terminal {
                 Json::array unitRaw = unitObj.array_items();
 
                 UNIT_TYPE unitType = static_cast<UNIT_TYPE>(i);
-                int x = unitRaw.at(0).int_value();
-                int y = unitRaw.at(1).int_value();
+                unsigned int x = unitRaw.at(0).int_value();
+                unsigned int y = unitRaw.at(1).int_value();
                 double hp = unitRaw.at(2).number_value();
 
                 gameMap.addUnit(unitType, x, y, player.id, hp);
@@ -211,7 +211,7 @@ namespace terminal {
     /// @param y The y position to check.
     /// @param num The number of units to check, default is 1.
     /// @return A bool, true if can spawn.
-    bool GameState::canSpawn(UNIT_TYPE unitType, int x, int y, int num) const {
+    bool GameState::canSpawn(UNIT_TYPE unitType, unsigned int x, unsigned int y, int num) const {
         if (!gameMap.inArenaBounds(x, y)) {
             Util::printError<PosException>("Out of bounds exception", CRASH, verbosity);
             return false;
@@ -261,7 +261,7 @@ namespace terminal {
     /// @param pos The location to spawn the unit.
     /// @param num The number of units to spawn.
     /// @return Returns the number of units spawned (0 or 1).
-    unsigned int GameState::attemptSpawn(UNIT_TYPE unitType, int x, int y, int num) {
+    unsigned int GameState::attemptSpawn(UNIT_TYPE unitType, unsigned int x, unsigned int y, int num) {
         if (canSpawn(unitType, x, y)) {
             int cost = (int)typeCost(unitType);
             RESOURCE resourceType = resourceRequired(unitType);
@@ -270,9 +270,9 @@ namespace terminal {
             gameMap.addUnit(unitType, x, y, 0);
 
             if (isStationary(unitType))
-                buildStack.push_back({ Json::array({ unitTypeStr(unitType), x, y }) });
+                buildStack.push_back({ Json::array({ unitTypeStr(unitType), (int)x, (int)y }) });
             else
-                deployStack.push_back({ Json::array({ unitTypeStr(unitType), x, y }) });
+                deployStack.push_back({ Json::array({ unitTypeStr(unitType), (int)x, (int)y }) });
 
             return 1;
         }
@@ -314,10 +314,10 @@ namespace terminal {
     /// @param x The x location to try and remove.
     /// @param y The y location to try and remove.
     /// @return Returns the number of units removed (0 or 1).
-    unsigned int GameState::attemptRemove(int x, int y) {
+    unsigned int GameState::attemptRemove(unsigned int x, unsigned int y) {
         if (y < gameMap.HALF_ARENA &&
             gameMap.containsStationaryUnit(Pos(x, y))) {
-            buildStack.push_back({ Json::array({ unitTypeStr(REMOVE), x, y}) });
+            buildStack.push_back({ Json::array({ unitTypeStr(REMOVE), (int)x, (int)y}) });
 
             return 1;
         }
